creating the othe methods file
